move communicators into sdk-internals directory
