+++ sdk-internals/communicator/none/communicator.go
