+++ sdk-internals/communicator/none/communicator_test.go
