+++ sdk-internals/communicator/ssh/connect_test.go
