+++ sdk-internals/communicator/ssh/connection.go
