+++ sdk-internals/communicator/ssh/keyboard_interactive_test.go
