+++ sdk-internals/communicator/ssh/password.go
