+++ sdk-internals/communicator/ssh/password_test.go
