+++ sdk-internals/communicator/winrm/communicator_test.go
