+++ sdk-internals/communicator/winrm/config.go
